+++ src/provider_runtime/openai_compatible.py
@@ -38,7 +38,7 @@ class OpenAICompatibleChatClient:
         req: ModelCall,
         *,
         api_key: str,
-        timeout_s: int,
+        timeout_s: float,
     ) -> ModelResponse:
         headers = self._build_headers(api_key)
         body = self._build_request_body(req, stream=False)
@@ -64,7 +64,7 @@ class OpenAICompatibleChatClient:
         req: ModelCall,
         *,
         api_key: str,
-        timeout_s: int,
+        timeout_s: float,
     ) -> AsyncIterator[ModelChunk]:
         headers = self._build_headers(api_key)
         body = self._build_request_body(req, stream=True)
